requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
langgraph==0.0.26
langchain-openai==0.0.5
langchain-core==0.1.23
langsmith==0.0.87
python-dotenv==1.0.0
requests==2.31.0
httpx==0.26.0